5-lc-advanced/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
pinecone
langchain-pinecone
python-dotenv
pypdf
streamlit